last version

+++ libft/ft_strlen.c
@@ -6,16 +6,18 @@
 /*   By: cstate <cstate@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/18 17:24:04 by cstate            #+#    #+#             */
-/*   Updated: 2024/11/18 17:24:59 by cstate           ###   ########.fr       */
+/*   Updated: 2024/11/29 12:51:45 by cstate           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+#include "libft.h"
 
 /* 
 	<<ft_strlen>> returns the length of the string [str]
  */
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != 0)
